benchmark/requirements.txt
fpdf2==2.8.5
python-docx==1.2.0
pandas==2.2.3
langchain-ollama==1.0.0
ragas==0.4.1
numpy==2.3.5
pydantic==2.12.5